requirements.txt
streamlit
ccxt
pandas
pandas_ta
plotly
python-dotenv
requests
watchdog